requirements.txt
disnake
asyncpg
python-dotenv
